checklist changes

+++ src/main/java/com/SWOOSH/repository/ReviewRepository.java
@@ -3,5 +3,6 @@ package com.SWOOSH.repository;
 import com.SWOOSH.model.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+
 }
